+++ Code/Combat/control.cpp
@@ -163,7 +163,7 @@ enum {
 void	ControlClass::Import_Cs( BitStreamClass & packet )
 { 
 	// Or in the new one time bits
-	ULONG otb_bits = packet.Get(otb_bits, BITPACK_ONE_TIME_BOOLEAN_BITS);
+	uint32_t otb_bits = packet.Get(otb_bits, BITPACK_ONE_TIME_BOOLEAN_BITS);
 	OneTimeBooleanBits |= otb_bits;
 
 	packet.Get(ContinuousBooleanBits, BITPACK_CONTINUOUS_BOOLEAN_BITS);
@@ -400,14 +400,14 @@ float	ControlClass::Get_Clamp(AnalogControl control)
 */
 
 /*
-BYTE ControlClass::Scale_Analog(float clamp, float unscaled) 
+uint8_t ControlClass::Scale_Analog(float clamp, float unscaled) 
 {
    WWASSERT(unscaled > -clamp - WWMATH_EPSILON);
    WWASSERT(unscaled < +clamp + WWMATH_EPSILON);
-   return (BYTE) (255 * WWMath::Clamp((unscaled + clamp) / (2 * clamp)));
+   return (uint8_t) (255 * WWMath::Clamp((unscaled + clamp) / (2 * clamp)));
 }
 
-float ControlClass::Unscale_Analog(float clamp, BYTE scaled) 
+float ControlClass::Unscale_Analog(float clamp, uint8_t scaled) 
 {
    float unscaled;
    if (scaled == 127) { // scaling perturbs zero, special-case it
