+++ Code/Combat/raveshawbossgameobj.cpp
@@ -280,7 +280,7 @@ RaveshawBossGameObjDefClass::~RaveshawBossGameObjDefClass (void)
 //	Get_Class_ID
 //
 //////////////////////////////////////////////////////////////////////////
-uint32
+uint32_t
 RaveshawBossGameObjDefClass::Get_Class_ID (void) const	
 { 
 	return CLASSID_GAME_OBJECT_DEF_RAVESHAW_BOSS; 
@@ -1451,7 +1451,7 @@ RaveshawBossGameObjClass::STATE_IMPL_BEGIN(MOVE_STATE_GOTO_TIBERIUM) (void)
 	//	Make the mutant move to that position
 	//
 	ActionParamsStruct params;
-	params.Set_Basic ((long)0, 100, 777);
+	params.Set_Basic (static_cast<int32_t>(0), 100, 777);
 	params.Set_Movement (CurrentDestPos, 0.6F, 0.5F);
 	Get_Action ()->Goto (params);
 	return ;
@@ -1507,7 +1507,7 @@ RaveshawBossGameObjClass::STATE_IMPL_BEGIN(MOVE_STATE_GOTO_CATWALK) (void)
 	//	Make the mutant move into position
 	//
 	ActionParamsStruct params;
-	params.Set_Basic ((long)0, 100, 777);
+	params.Set_Basic (static_cast<int32_t>(0), 100, 777);
 	params.Set_Movement (CurrentDestPos, 0.6F, 0.5F);
 	Get_Action ()->Goto (params);
 	return ;
@@ -1689,7 +1689,7 @@ RaveshawBossGameObjClass::STATE_IMPL_THINK(MOVE_STATE_CIRCLE_CATWALK) (void)
 			//	Now, simply make the mutant move to this position
 			//
 			ActionParamsStruct params;
-			params.Set_Basic ((long)0, 100, 777);
+			params.Set_Basic (static_cast<int32_t>(0), 100, 777);
 			params.Set_Movement (CurrentDestPos, 0.5F, 0.5F);
 			params.IgnoreFacing = true;
 			Get_Action ()->Goto (params);		
@@ -1712,7 +1712,7 @@ RaveshawBossGameObjClass::STATE_IMPL_BEGIN(MOVE_STATE_GOTO_THROW_OBJECT) (void)
 	//	Walk to the position of the throw object
 	//
 	ActionParamsStruct params;
-	params.Set_Basic ((long)0, 100, 777);
+	params.Set_Basic (static_cast<int32_t>(0), 100, 777);
 	params.Set_Movement (ThrownObject, 0.5F, 1.25F);
 	Get_Action ()->Goto (params);		
 	return ;
@@ -1842,7 +1842,7 @@ RaveshawBossGameObjClass::STATE_IMPL_THINK(MOVE_STATE_FOLLOW_STAR) (void)
 			//	Track down the player
 			//
 			ActionParamsStruct params;
-			params.Set_Basic ((long)0, 100, 777);
+			params.Set_Basic (static_cast<int32_t>(0), 100, 777);
 			params.Set_Movement (COMBAT_STAR, 1.0F, 1.0F);
 			params.MoveFollow		= true;
 			Get_Action ()->Goto (params);
